refactoring: one of "Rectangle" methods move to  "used_frequently"
