+++ Task 1/Navigation_bar/THE_MAGNIFICENT_TEN.csproj
@@ -9,6 +9,8 @@
 
   <ItemGroup>
     <ProjectReference Include="..\1.1.1_Rectangle\1.1.1_Rectangle.csproj" />
+    <ProjectReference Include="..\1.1.2_Triangle\1.1.2_Triangle.csproj" />
+    <ProjectReference Include="..\1.1.3_Another_triaangle\1.1.3_Another_triangle.csproj" />
   </ItemGroup>
 
 </Project>
