+++ Task 1/used_frequently/used_frequently.csproj
@@ -2,13 +2,8 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>_1._1._1_Rectangle</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\used_frequently\used_frequently.csproj" />
-  </ItemGroup>
 
 </Project>
